Tutorial 40: BallGame, Variadic Macros, SRP

+++ Advanced C++ Graphics T Pt 2/Gun.cpp
@@ -41,7 +41,7 @@ void Gun::update(bool isMouseDown, const glm::vec2& position, const glm::vec2& d
 
 void Gun::fire(const glm::vec2& direction, const glm::vec2& position, std::vector<Bullet>& bullets) {
 
-	static std::mt19937 randomEngine(time(nullptr));
+	static std::mt19937 randomEngine((unsigned int)time(nullptr));
 	std::uniform_real_distribution<float> randRotate(-m_spread, m_spread);
 
 	m_fireEffect.play();
